+++ Assets/Scripts/Generics/RandomElement.cs
@@ -4,13 +4,13 @@
 // Avoid infinite loop and to have the same value twice in a row
 public class RandomElement
 {
-	private const int MAX_RANDOM_ITERATION = 3;                     // Max number to iterate in a list to avoid a infinite loop
+	private const ushort MAX_RANDOM_ITERATION = 3;                     // Max number to iterate in a list to avoid a infinite loop
 
-	public int GetCurrentIndex => _currentIndex;
+	public uint GetCurrentIndex => _currentIndex;
 
-	private int _currentIndex = 0;
+	private uint _currentIndex = 0;
 
-	public RandomElement(int currentIndex = 0) => _currentIndex = currentIndex;
+	public RandomElement(uint currentIndex = 0) => _currentIndex = currentIndex;
 
 	// Get an element int the list
 	public T Choose<T>(T[] elements)
@@ -26,13 +26,12 @@ public class RandomElement
 		// To obtain a different target that the current and to avoid infinite loop
 		while (i == _currentIndex && y < MAX_RANDOM_ITERATION);
 
-		_currentIndex = i;
-
+		_currentIndex = (uint)i;
 		return elements[_currentIndex];
 	}
 
 	// Get the index
-	public int Choose(int sizeElements)
+	public uint Choose(int sizeElements)
 	{
 		int i;
 		int y = 0;
@@ -45,7 +44,7 @@ public class RandomElement
 		// To obtain a different target that the current and to avoid infinite loop
 		while (i == _currentIndex && y < MAX_RANDOM_ITERATION);
 
-		_currentIndex = i;
+		_currentIndex = (uint)i;
 		return _currentIndex;
 	}
 }
