+++ Assets/Scripts/Scores/ScoreCollidable.cs
@@ -3,7 +3,7 @@
 // Increase score if it collides with something
 public class ScoreCollidable : MonoBehaviour, ICollidable
 {
-	[SerializeField] private int _addScore = 10;
+	[SerializeField] private ushort _addScore = 10;
 
 	public void Hit() => ScoreManager.Instance.BulletModifier += _addScore;
 }
